Update current_results.py
Have reversed the print order for the 'list all' function so the latest result appears at the bottom. This is done by compiling a list of all results played using timestamps & numerical switches to separate & collect data. list.reverse() is applied so the data is printed in descending order. The script then uses 'tabulate' and 'zip_longest' to append each string in the list with a new line.