Update run.py
A new process has been added to the update functions that will check for the presence of the geckodriver.exe processes the program uses to obtain its updates, when the program is detected, it is killed. The function is ran at the end of the update process itself so the application can automatically fully close. The path of geckodriver is additionally analysed and the software is only killed when the path matches the extraction path of the program, this way, the application doesnt close other geckodriver processes the user might be running.