New files for check & delete geckodriver processes
Two new files used in the check & delete process. get_gecko_path is used to determine whether the running geckodriver.exe is running from the extraction directory of the main program itself. If it matches, it uses get_gecko_pid to query the PID of the process, using this to kill the task with SIGTERM.