Update run_update.py
Changes have been made to the mthod in which the external version number reference is obtained for the update process. Now, instead of reading the number from an update table, the update program now reads the version numbers directly from the files online (first line) and uses that for comparison. The same function is performed with the local files stored on the users end, in this manner, the version numbers have all been correctly embedded into the data files themselves! (#25) (#15)